Move unit builder to its own class
